Omit empty <lastmod> (#70)
If `lastmod` is omitted in `additions` in the YAML config it’s still included as an empty element in the XML output. Empty values for `<lastmod>` are invalid and result in errors in the Google Search Console.

The `<lastmod>` element is [optional](https://www.google.com/sitemaps/protocol.html) and should be omitted if it’s empty.